💄 마이페이지 최대 폭 698px로 설정
