+++ src/pages/CompanyMyPage.tsx
@@ -33,7 +33,7 @@ export const CompanyMyPage = () => {
     <div className="min-h-screen bg-grey-50">
       <GlobalNavigationBar />
       {/* 메인 컨텐츠 */}
-      <div className="mx-auto flex w-full flex-col gap-10 px-6 py-[30px] sm:w-screen-sm md:w-screen-md lg:w-screen-lg xl:max-w-screen-xl">
+      <div className="w-max-[698px] mx-auto flex flex-col gap-10 px-6 py-[30px]">
         <h1 className="text-2xl font-bold text-grey-900">마이페이지</h1>
         <Tabs
           value={
